replaced than and chose with appropriate versions
"Chose" is past tense, "choose" is present and future tense. From the context it is clear that the latter should be preferred.

"Than" is a comparative conjunction, used in instances such as 'greater than', 'less than', etc. "Then" is used to indicate actions or progressions in time (e.g is a sequence of events).
The ternary operators in languages such as Java are merely terse ways of writing "if then else", thus "then" is the appropriate word to use.